make add to cart function reusable
